an improved version
first, fixed the printing to be more useful.
added some more usecases.
Added code to fix cyclic deps.
add swap_vals to swap_vals in case we try to
push something that was visited.
it means we already added it but it
(the values that we are looking at right now)
 needs to come before